# docs/offeredapis/swagger/pms-api.yaml
openapi: 3.0.1
info:
  title: A1 Policy management service
  description: <h2>General</h2><p>The O-RAN Non-RT RIC Policy Management Service provides
    a REST API for management of A1 policices. <br/>The main tasks of the service
    are:</p><ul><li>A1 Policy creation, modification and deletion.</li><li>Monitoring
    and maintaining consistency of the SMO view of A1 policies and the Near-RT RICs</li><li>Maintaining
    a view of supported Near-RT RIC policy types</li><li>Supervision of using services
    (R-APPs). When a service is unavailble, its policies are removed.</li></ul><h2>APIs
    provided by the service</h2><h3>A1 Policy Management</h3><p>This is an API for
    management of A1 Policies.</p><ul><li>A1 Policy retrieval, creation, modification
    and deletion.</li><li>Retrieval of supported A1 Policy types for a Near-RT RIC</li><li>Retrieval
    of status for existing A1 policies</li></ul><h3>Management of configuration</h3><p>API
    for updating and retrieval of the component configuration. Note that there other
    ways to maintain the configuration.</p><h3>Callbacks</h3><p>These are endpoints
    that are invoked by this service. The callbacks are registerred in this service
    at service registration.</p><h3>NearRT-RIC Repository</h3><p>This is an API that
    provides support for looking up a NearRT-RIC. Each A1 policy is targeted for one
    Near-RT RIC.</p><h3>Health Check</h3><p>API used for supervision of the PMS component.</p><h3>Service
...
  license:
...
servers:
...
paths:
...
  /a1-policy/v2/configuration:
    get:
      tags:
      - Management of configuration
      summary: Returns the contents of the configuration file
      description: Note that the file contents is not relevant if the Consul is used.
      operationId: getConfiguration
      responses:
        200:
          description: Configuration
          content:
            application/json:
              schema:
                type: object
        404:
          description: File is not found or readable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
    put:
      tags:
      - Management of configuration
      summary: Replace the current configuration file with the given configuration
      description: Note that the file is ignored if the Consul is used.
      operationId: putConfiguration
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: Configuration updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        400:
          description: Invalid configuration provided
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
        500:
          description: Something went wrong when replacing the configuration. Try
            again.
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
...
components:
  schemas:
    error_information:
      type: object
      properties:
        detail:
          type: string
          description: ' A human-readable explanation specific to this occurrence
            of the problem.'
          example: Policy type not found
        status:
          type: integer
          description: 'The HTTP status code generated by the origin server for this
            occurrence of the problem. '
          format: int32
          example: 503
      description: Problem as defined in https://tools.ietf.org/html/rfc7807
    void:
      type: object
      description: Void/empty
...